feat(api): move copy target profile stages to evaluation context
