+++ api/src/prescription/target-profile/application/admin-target-profile-controller.js
@@ -1,6 +1,5 @@
 import dayjs from 'dayjs';
 
-import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { evaluationUsecases } from '../../../evaluation/domain/usecases/index.js';
 import * as organizationSerializer from '../../../organizational-entities/infrastructure/serializers/jsonapi/organization-serializer.js';
 import { withTransaction } from '../../../shared/domain/DomainTransaction.js';
@@ -163,7 +162,7 @@ const copyTargetProfile = withTransaction(async (request) => {
       originTargetProfileId: targetProfileIdToCopy,
       destinationTargetProfileId: copiedTargetProfileId,
     }),
-    await usecases.copyTargetProfileStages({
+    await evaluationUsecases.copyTargetProfileStages({
       originTargetProfileId: targetProfileIdToCopy,
       destinationTargetProfileId: copiedTargetProfileId,
     }),
